+++ RingSoft.DevLogix.DataAccess/Model/ProjectManagement/LaborPart.cs
@@ -14,7 +14,7 @@ namespace RingSoft.DevLogix.DataAccess.Model.ProjectManagement
         public string Name { get; set; }
 
         [Required]
-        public decimal MinutesCost { get; set; }
+        public double MinutesCost { get; set; }
 
         public string? Comment { get; set; }
