+++ RingSoft.DevLogix.DataAccess/Model/ProjectManagement/ProjectMaterialPart.cs
@@ -31,8 +31,8 @@ namespace RingSoft.DevLogix.DataAccess.Model.ProjectManagement
         [MaxLength(50)]
         public string? Description { get; set; }
 
-        public decimal? Quantity { get; set; }
+        public double? Quantity { get; set; }
 
-        public decimal? Cost { get; set; }
+        public double? Cost { get; set; }
     }
 }
